6-size.c

+++ 0x00-hello_world/6-size.c
@@ -8,11 +8,11 @@ int main(void)
 	long d;
 	long long e;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned int)sizeof(a));
-	printf("Size of an int: %lu byte(s)\n", (unsigned int)sizeof(b));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(e));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("Size of a float: %lu byte(s)\n", (unsigned int)sizeof(c));
+	printf("Size of a char: %lu byte(s)\n", sizeof(a));
+	printf("Size of an int: %lu byte(s)\n", sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n", sizeof (e));
+	printf("Size of a long long int: %lu byte(s)\n", sizeof(d));
+	printf("Size of a float: %lu byte(s)\n", sizeof(c));
 
 	return (0);
 }
